Refactoring Tests
